Split status rendering out of the dropbox command module
commands/sync.rs had grown past 800 lines holding two unrelated jobs: driving
the transfer commands and rendering the local-vs-remote status table. The
table moves to commands/sync_status.rs, leaving each file around 400 lines.

`format_size` existed twice, in commands/sync.rs and commands/info.rs, with
identical behaviour. Both now use one copy in output::format, where the other
display helpers live.

No behaviour change; `dropbox status` output is unchanged.
